Add files via upload
Tourists and even general people travel from one place to
another in order to explore or for employment purposes but
before they explore, they want to get an insight data about the
place. So, to help them with this, a simple city project can be
the best guide for them. It is a web-based application written in
Java which basically guides you about the place you’re going to
visit. You can access all the details of the city. In this
application, users need to sign up by entering input details and
then can access all the required details of the city. It contains
various modules like admin, tourism, business, and student
wherein users can switch to the module as per the requirement.